refarctoring + change comments+typing
