+++ storey/__init__.py
@@ -10,15 +10,18 @@ from .drivers import (  # noqa: F401
     V3ioDriver, NoopDriver
 )
 from .dtypes import (  # noqa: F401
-    Event, FieldAggregator, Table, Context
+    Event, FieldAggregator
 )
 from .flow import (  # noqa: F401
     Filter, FlatMap, Flow, FlowError, JoinWithV3IOTable, SendToHttp, JoinWithTable, Map, Extend, Reduce, Batch,
-    MapWithState, MapClass, Complete, Choice, HttpRequest, HttpResponse, build_flow
+    MapWithState, MapClass, Complete, Choice, HttpRequest, HttpResponse, build_flow, Context
 )
 from .sources import (  # noqa: F401
     Source, AsyncSource, DataframeSource, ReadCSV
 )
 from .writers import (  # noqa: F401
     WriteToV3IOStream, WriteToCSV, WriteToParquet, WriteToTSDB, WriteToTable
+)
+from .table import (  # noqa: F401
+    Table
 )
