+++ storey/writers.py
@@ -9,8 +9,9 @@ import pandas as pd
 import v3io_frames as frames
 
 from . import V3ioDriver
-from .dtypes import V3ioError, Event, Table
+from .dtypes import V3ioError, Event
 from .flow import Flow, _termination_obj, _split_path, _Batching, _ConcurrentByKeyJobExecution
+from .table import Table
 
 
 class _Writer:
@@ -381,8 +382,7 @@ class WriteToTable(_ConcurrentByKeyJobExecution, _Writer):
     """
     Persists the data in `table` to its associated storage by key.
 
-    :param table: A table object. Alternatively can provide the table's name if a the relevant Table entry exists
-     in the context provided.
+    :param table: A Table object or name to persist. If a table name is provided, it will be looked up in the context.
     :param columns: Fields to be written to the storage. Will be extracted from events when an event is a dictionary (lists will be written
     as is). Use = notation for renaming fields (e.g. write_this=event_field).
     Use $ notation to refer to metadata ($key, event_time=$time). Optional. Defaults to None (will be inferred if event is dictionary).
@@ -397,7 +397,7 @@ class WriteToTable(_ConcurrentByKeyJobExecution, _Writer):
         self._table = table
         if isinstance(table, str):
             if not self.context:
-                raise TypeError("table can not be string if no context was provided to the step")
+                raise TypeError("Table can not be string if no context was provided to the step")
             self._table = self.context.get_table(table)
         self._closeables = [self._table]
